removed dedicated recipe, realized it could be generalized
