+++ src/test/java/org/openrewrite/java/testing/hamcrest/HamcrestMatcherToAssertJTest.java
@@ -366,6 +366,40 @@ class HamcrestMatcherToAssertJTest implements RewriteTest {
                   """)
             );
         }
+
+        @Test
+        void CloseToTest() {
+            rewriteRun(
+              spec -> spec.recipe(new HamcrestMatcherToAssertJ("closeTo", "isCloseTo")),
+              //language=java
+              java("""
+                  import org.junit.jupiter.api.Test;
+                  
+                  import static org.hamcrest.MatcherAssert.assertThat;
+                  import static org.hamcrest.Matchers.closeTo;
+                  
+                  class ATest {
+                      @Test
+                      void replaceCloseTo() {
+                          assertThat(1.0, closeTo(2.0, 1.0));
+                      }
+                  }
+                  """,
+                """
+                  import org.junit.jupiter.api.Test;
+                  
+                  import static org.assertj.core.api.Assertions.assertThat;
+                  import static org.assertj.core.api.Assertions.within;
+                  
+                  class ATest {
+                      @Test
+                      void replaceCloseTo() {
+                          assertThat(1.0).isCloseTo(2.0, within(1.0));
+                      }
+                  }
+                  """)
+            );
+        }
     }
 
     @Nested
